refactor: Move helpers from onOperationOrBillCreate
